# Replace LongDecimalWithOverflowAndLongState for decimal operators