--- a/core/trino-main/src/main/java/io/trino/operator/aggregation/DecimalSumAggregation.java
+++ b/core/trino-main/src/main/java/io/trino/operator/aggregation/DecimalSumAggregation.java
@@ -21,9 +21,11 @@ import io.trino.metadata.FunctionNullability;
 import io.trino.metadata.Signature;
 import io.trino.metadata.SqlAggregationFunction;
 import io.trino.operator.aggregation.AggregationMetadata.AccumulatorStateDescriptor;
-import io.trino.operator.aggregation.state.LongDecimalWithOverflowState;
-import io.trino.operator.aggregation.state.LongDecimalWithOverflowStateFactory;
-import io.trino.operator.aggregation.state.LongDecimalWithOverflowStateSerializer;
+import io.trino.operator.aggregation.state.Int128State;
+import io.trino.operator.aggregation.state.Int128StateFactory;
+import io.trino.operator.aggregation.state.Int128StateSerializer;
+import io.trino.operator.aggregation.state.NullableLongState;
+import io.trino.operator.aggregation.state.StateCompiler;
 import io.trino.spi.block.Block;
 import io.trino.spi.block.BlockBuilder;
 import io.trino.spi.type.DecimalType;
@@ -49,12 +51,11 @@ public class DecimalSumAggregation
 {
     public static final DecimalSumAggregation DECIMAL_SUM_AGGREGATION = new DecimalSumAggregation();
     private static final String NAME = "sum";
-    private static final MethodHandle SHORT_DECIMAL_INPUT_FUNCTION = methodHandle(DecimalSumAggregation.class, "inputShortDecimal", LongDecimalWithOverflowState.class, Block.class, int.class);
-    private static final MethodHandle LONG_DECIMAL_INPUT_FUNCTION = methodHandle(DecimalSumAggregation.class, "inputLongDecimal", LongDecimalWithOverflowState.class, Block.class, int.class);
+    private static final MethodHandle SHORT_DECIMAL_INPUT_FUNCTION = methodHandle(DecimalSumAggregation.class, "inputShortDecimal", Int128State.class, NullableLongState.class, Block.class, int.class);
+    private static final MethodHandle LONG_DECIMAL_INPUT_FUNCTION = methodHandle(DecimalSumAggregation.class, "inputLongDecimal", Int128State.class, NullableLongState.class, Block.class, int.class);
+    private static final MethodHandle LONG_DECIMAL_OUTPUT_FUNCTION = methodHandle(DecimalSumAggregation.class, "outputLongDecimal", Int128State.class, NullableLongState.class, BlockBuilder.class);
 
-    private static final MethodHandle LONG_DECIMAL_OUTPUT_FUNCTION = methodHandle(DecimalSumAggregation.class, "outputLongDecimal", LongDecimalWithOverflowState.class, BlockBuilder.class);
-
-    private static final MethodHandle COMBINE_FUNCTION = methodHandle(DecimalSumAggregation.class, "combine", LongDecimalWithOverflowState.class, LongDecimalWithOverflowState.class);
+    private static final MethodHandle COMBINE_FUNCTION = methodHandle(DecimalSumAggregation.class, "combine", Int128State.class, NullableLongState.class, Int128State.class, NullableLongState.class);
 
     public DecimalSumAggregation()
     {
@@ -80,8 +81,6 @@ public class DecimalSumAggregation
         Type inputType = getOnlyElement(boundSignature.getArgumentTypes());
         checkArgument(inputType instanceof DecimalType, "type must be Decimal");
         MethodHandle inputFunction;
-        Class<LongDecimalWithOverflowState> stateInterface = LongDecimalWithOverflowState.class;
-        LongDecimalWithOverflowStateSerializer stateSerializer = new LongDecimalWithOverflowStateSerializer();
 
         if (((DecimalType) inputType).isShort()) {
             inputFunction = SHORT_DECIMAL_INPUT_FUNCTION;
@@ -95,91 +94,105 @@ public class DecimalSumAggregation
                 Optional.empty(),
                 Optional.of(COMBINE_FUNCTION),
                 LONG_DECIMAL_OUTPUT_FUNCTION,
-                ImmutableList.of(new AccumulatorStateDescriptor<>(
-                        stateInterface,
-                        stateSerializer,
-                        new LongDecimalWithOverflowStateFactory())));
+                ImmutableList.of(
+                        new AccumulatorStateDescriptor<>(
+                                Int128State.class,
+                                new Int128StateSerializer(),
+                                new Int128StateFactory()),
+                        new AccumulatorStateDescriptor<>(
+                                NullableLongState.class,
+                            StateCompiler.generateStateSerializer(NullableLongState.class),
+                            StateCompiler.generateStateFactory(NullableLongState.class))));
     }
 
-    public static void inputShortDecimal(LongDecimalWithOverflowState state, Block block, int position)
+    public static void inputShortDecimal(Int128State decimalState, NullableLongState overflowState, Block block, int position)
     {
-        state.setNotNull();
+        long[] decimal = decimalState.getArray();
+        int decimalOffset = decimalState.getArrayOffset();
 
-        long[] decimal = state.getDecimalArray();
-        int offset = state.getDecimalArrayOffset();
-
+        decimalState.setNotNull();
         long rightLow = block.getLong(position, 0);
         long rightHigh = rightLow >> 63;
 
         long overflow = addWithOverflow(
-                decimal[offset],
-                decimal[offset + 1],
+                decimal[decimalOffset],
+                decimal[decimalOffset + 1],
                 rightHigh,
                 rightLow,
                 decimal,
-                offset);
-        state.setOverflow(Math.addExact(overflow, state.getOverflow()));
+                decimalOffset);
+        overflowState.setValue(Math.addExact(overflow, overflowState.getValue()));
+        overflowState.setNull(false);
     }
 
-    public static void inputLongDecimal(LongDecimalWithOverflowState state, Block block, int position)
+    public static void inputLongDecimal(Int128State decimalState, NullableLongState overflowState, Block block, int position)
     {
-        state.setNotNull();
+        long[] decimal = decimalState.getArray();
+        int decimalOffset = decimalState.getArrayOffset();
 
-        long[] decimal = state.getDecimalArray();
-        int offset = state.getDecimalArrayOffset();
-
+        decimalState.setNotNull();
         long rightHigh = block.getLong(position, 0);
         long rightLow = block.getLong(position, SIZE_OF_LONG);
 
         long overflow = addWithOverflow(
-                decimal[offset],
-                decimal[offset + 1],
+                decimal[decimalOffset],
+                decimal[decimalOffset + 1],
                 rightHigh,
                 rightLow,
                 decimal,
-                offset);
+                decimalOffset);
 
-        state.addOverflow(overflow);
+        overflowState.setValue(overflowState.getValue() + overflow);
+        overflowState.setNull(false);
     }
 
-    public static void combine(LongDecimalWithOverflowState state, LongDecimalWithOverflowState otherState)
+    public static void combine(Int128State decimalState, NullableLongState overflowState, Int128State otherDecimalState, NullableLongState otherOverflowState)
     {
-        long[] decimal = state.getDecimalArray();
-        int offset = state.getDecimalArrayOffset();
+        if (!decimalState.isNotNull() && !otherDecimalState.isNotNull()) {
+            return;
+        }
 
-        long[] otherDecimal = otherState.getDecimalArray();
-        int otherOffset = otherState.getDecimalArrayOffset();
+        long[] decimal = decimalState.getArray();
+        int decimalOffset = decimalState.getArrayOffset();
 
-        if (state.isNotNull()) {
+        long[] otherDecimal = otherDecimalState.getArray();
+        int otherDecimalOffset = otherDecimalState.getArrayOffset();
+
+        if (decimalState.isNotNull()) {
             long overflow = addWithOverflow(
-                    decimal[offset],
-                    decimal[offset + 1],
-                    otherDecimal[otherOffset],
-                    otherDecimal[otherOffset + 1],
+                    decimal[decimalOffset],
+                    decimal[decimalOffset + 1],
+                    otherDecimal[otherDecimalOffset],
+                    otherDecimal[otherDecimalOffset + 1],
                     decimal,
-                    offset);
-            state.addOverflow(Math.addExact(overflow, otherState.getOverflow()));
+                    decimalOffset);
+            overflowState.setValue(overflowState.getValue() + Math.addExact(overflow, otherOverflowState.getValue()));
+            overflowState.setNull(false);
         }
         else {
-            state.setNotNull();
-            decimal[offset] = otherDecimal[otherOffset];
-            decimal[offset + 1] = otherDecimal[otherOffset + 1];
-            state.setOverflow(otherState.getOverflow());
+            // TODO: Dlaczego tego teraz potrzebujemy a kiedys nie?
+            if (!otherDecimalState.isNotNull()) {
+                return;
+            }
+            decimalState.setNotNull();
+            decimal[decimalOffset] = otherDecimal[otherDecimalOffset];
+            decimal[decimalOffset + 1] = otherDecimal[otherDecimalOffset + 1];
+            overflowState.set(otherOverflowState);
         }
     }
 
-    public static void outputLongDecimal(LongDecimalWithOverflowState state, BlockBuilder out)
+    public static void outputLongDecimal(Int128State decimalState, NullableLongState overflowState, BlockBuilder out)
     {
-        if (state.isNotNull()) {
-            if (state.getOverflow() != 0) {
+        if (decimalState.isNotNull()) {
+            if (overflowState.getValue() != 0) {
                 throw new ArithmeticException("Decimal overflow");
             }
 
-            long[] decimal = state.getDecimalArray();
-            int offset = state.getDecimalArrayOffset();
+            long[] decimal = decimalState.getArray();
+            int decimalOffset = decimalState.getArrayOffset();
 
-            long rawHigh = decimal[offset];
-            long rawLow = decimal[offset + 1];
+            long rawHigh = decimal[decimalOffset];
+            long rawLow = decimal[decimalOffset + 1];
 
             Decimals.throwIfOverflows(rawHigh, rawLow);
             out.writeLong(rawHigh);