… core/trino-main/src/main/java/io/trino/operator/aggregation/state/Int128StateFactory.java
… core/trino-main/src/main/java/io/trino/operator/aggregation/state/LongDecimalWithOverflowAndLongStateFactory.java
… core/trino-main/src/main/java/io/trino/operator/aggregation/state/LongDecimalWithOverflowAndLongStateSerializer.java
… core/trino-main/src/main/java/io/trino/operator/aggregation/state/LongDecimalWithOverflowState.java
… core/trino-main/src/main/java/io/trino/operator/aggregation/state/LongDecimalWithOverflowStateFactory.java
… core/trino-main/src/main/java/io/trino/operator/aggregation/state/LongDecimalWithOverflowStateSerializer.java
… core/trino-main/src/test/java/io/trino/operator/aggregation/TestDecimalAverageAggregation.java
… core/trino-main/src/test/java/io/trino/operator/aggregation/TestDecimalSumAggregation.java